+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix request param casting (array to string)
 - Replace TinyMCE with Tiptap
 - Fix verbose required field error messages
+- Fix locale fields projection in content api
 
 ## 2.9.4 (2024-10-07)